+++ sql_restore_check_db/Auto_Restore_DB_ONLY.ps1
@@ -1,6 +1,8 @@
 ﻿######################################################## RESTORE DATABASE And CHECK DB ##############################################################################################
 $NameDB="PSE"
 
+$rezult_FO = $PSScriptRoot+"\PHYSICAL_ONLY_rezult.txt"
+$rezult_EL = $PSScriptRoot+"\EXTENDED_LOGICAL_rezult.txt"
 $tmp_f = $PSScriptRoot+"\tmp.txt"
 $rest_db = $PSScriptRoot+"\restore_db.sql"
 $rezult = $PSScriptRoot+"\RESTORE_SQL_REZULT.txt"
@@ -9,7 +11,12 @@ $path_backup_files_Z="Z:\$NameDB\"
 $path_backup_files_W="W:\$NameDB\"
 $ErrorActionPreference= "Continue"
 $Error.Clear()
-Write-Output "The value is[$Error]"
+
+
+$query_state_db="SELECT state_desc
+    FROM sys.databases
+    WHERE Name = '$NameDB'
+    GO"
 
 #-------------------------------------------------------Function "WriteLog"---------------------------------------------------------------------------------------------------------
 $Logfile = $PSScriptRoot+"\DB_RESTORE_CHECKDB_LOG.txt"
@@ -68,27 +75,19 @@ if ((test-path $path_backup_files_Y) -and (test-path $path_backup_files_Z))
                 $time_res_stop = (Get-Date)
                 $time_res_total=($time_res_stop-$time_res_start).ToString().Split('.')[0]
 
-
-
-                $query_state_db="SELECT state_desc
-                FROM    sys.databases
-                WHERE Name = '$NameDB'
-                GO"
+                #-------------------------------------------------------CHECKING RESTORED DATABASE + TRANSACTION LOG CORRECTLY------------------------------------------------------------------------
 
                 $db_status=Invoke-Sqlcmd -ServerInstance localhost -Database master -SuppressProviderContextWarning $query_state_db
-                #$db_status=($db_status[0] | Out-String).Trim()
-
                 if (($db_status[0] | Out-String).Trim() -eq "ONLINE")
-
                 {
-                "БД восстановлена, в состоянии: $db_status"
+                Write-Output "БД восстановлена, в состоянии: $db_status"
                 'Время восстановления БД + цепочки файлов транзакций : '+ $time_res_total| out-file -Filepath $rezult -append
                 Remove-Item $rest_db
                 }
 
                 else
                 {
-                "БД не восстановлена, в состоянии: $db_status"
+                Write-Output "БД не восстановлена, в состоянии: $db_status"
                 "БД не восстановлена, в состоянии: $db_status"| out-file -Filepath $rezult -append
                 Remove-Item $rest_db
                 exit
@@ -110,11 +109,27 @@ if ((test-path $path_backup_files_Y) -and (test-path $path_backup_files_Z))
                 #-------------------------------------------------------RESTORE DATABASE WITHOUT TRANSACTION LOG--------------------------------------------------------------------------------------------
 
                 $time_res_start = (Get-Date)
-                Invoke-Sqlcmd -ServerInstance localhost -Querytimeout 0 -Database "master" -InputFile $rest_db -Verbose 4>&1 | out-file -Filepath $rezult
+                Invoke-Sqlcmd -ServerInstance localhost -Querytimeout 0 -Database "master" -InputFile $rest_db -Verbose 2>&1 4>&1 | out-file -Filepath $rezult
                 $time_res_stop = (Get-Date)
                 $time_res_total=($time_res_stop-$time_res_start).ToString().Split('.')[0]
-                "`n"+'Время восстановления БД : '+ $time_res_total| out-file -Filepath $rezult -append
+
+                #-------------------------------------------------------CHECKING RESTORED DATABASE ONLY CORRECTLY------------------------------------------------------------------------
+
+                $db_status=Invoke-Sqlcmd -ServerInstance localhost -Database master -SuppressProviderContextWarning $query_state_db
+                if (($db_status[0] | Out-String).Trim() -eq "ONLINE")
+                {
+                Write-Output "БД восстановлена, в состоянии: $db_status"
+                'Время восстановления БД + цепочки файлов транзакций : '+ $time_res_total| out-file -Filepath $rezult -append
                 Remove-Item $rest_db
+                }
+
+                else
+                {
+                Write-Output "БД не восстановлена, в состоянии: $db_status"
+                "БД не восстановлена, в состоянии: $db_status"| out-file -Filepath $rezult -append
+                Remove-Item $rest_db
+                exit
+                }
        }#3_1
 
     }#2_1
@@ -145,7 +160,7 @@ else
     "----------------------------------------------------------"| out-file -Filepath $Logfile -append
 }
 #-------------------------------------------------------DELETE TEMP LOG--------------------------------------------------------------------------------------------------
-del $rezult
+Remove-Item $rezult
 
 
 #-------------------------------------------------------DELETE TEMP LOG FILES IF ERROR--------------------------------------------------------------------------------------------------
@@ -156,15 +171,15 @@ WriteLog "Удаление временных файлов завершено - 
 "----------------------------------------------------------" | out-file -Filepath $Logfile -append
     if (test-path $rezult_EL)
     {
-    del $rezult_EL
+    Remove-Item $rezult_EL
     }
     if (test-path $rezult_FO)
     {
-    del $rezult_FO
+    Remove-Item $rezult_FO
     }
     if (test-path $rezult)
     {
-    del $rezult
+    Remove-Item $rezult
     }
 }
 else
